Add new registry functions to do basic coordinate resizing.
This is intended to allow resizing of the game window.  To minimize
the work involved, we leave all dimensions in the source for a 800x500
window, and then call the adjwidth/adjheight/adjpos/adjrect functions
to get the correct dimensions.

In driver.py, resize image assets to the appropriate size.  Also,
create new resource 'rsprites' to contain the reversed version of the
sprites, instead of flipping them every frame.

In all other functions, use the resize functions appropriately.
Change the duck icon width to 18 to match the source.